requirements.txt
setuptools==50.3.2
wheel==0.35.1
twine==3.2.0
PyYAML~=5.3.1